requirements.txt
gunicorn==21.2.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.14.0
Werkzeug==3.1.3
pymongo==4.6.1
python-dotenv==1.0.0
redis==5.0.1
firebase-admin==6.4.0
cloudinary==1.39.0



PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
pandas>=1.5.0
openpyxl>=3.0.0
opencv-python>=4.5.0
python-dateutil>=2.8.0
regex>=2022.0.0


